refactor(backend): split openaiService into boundary-focused modules
Keep openaiService.ts as a compatibility facade while moving implementation into focused modules for request execution, metadata assembly, citation extraction, and shared types.\n\nPreserves existing exports and behavior, and adds explicit ownership comments at module boundaries to prevent new junk-drawer files.
