+++ packages/backend/src/services/openaiService/metadata.ts
@@ -1,0 +1,526 @@
+/**
+ * @description: Canonical backend response metadata assembly from assistant
+ * output metadata plus runtime execution context.
+ * @footnote-scope: utility
+ * @footnote-module: OpenAIServiceMetadata
+ * @footnote-risk: high - Metadata mistakes can misclassify provenance, TRACE chips, and execution timeline.
+ * @footnote-ethics: high - Users depend on this metadata for transparency and governance trust.
+ */
+
+import crypto from 'node:crypto';
+import type {
+    ExecutionEvent,
+    ExecutionReasonCode,
+    ExecutionStatus,
+    EvaluatorExecutionReasonCode,
+    GenerationExecutionReasonCode,
+    PlannerExecutionApplyOutcome,
+    PlannerExecutionContractType,
+    PlannerExecutionPurpose,
+    PlannerExecutionReasonCode,
+    ResponseMetadata,
+    SafetyTier,
+    SteerabilityControlId,
+    ToolInvocationReasonCode,
+    TraceAxisScore,
+} from '@footnote/contracts/ethics-core';
+import { runtimeConfig } from '../../config.js';
+import { logger } from '../../utils/logger.js';
+import {
+    classifyProvenanceWithSignals,
+    deriveRetrievedChips,
+    resolveTradeoffCount,
+} from '../responseMetadataHeuristics.js';
+import type {
+    AssistantResponseMetadata,
+    ResponseMetadataRuntimeContext,
+} from './types.js';
+
+// Owns: response metadata assembly and normalization of execution metadata fields.
+// Does not own: provider request execution or orchestrator policy decisions.
+
+const TRACE_AXIS_KEYS = [
+    'tightness',
+    'rationale',
+    'attribution',
+    'caution',
+    'extent',
+] as const;
+
+/**
+ * Runtime guard for TRACE axis chip values.
+ */
+const isTraceAxisScore = (value: unknown): value is TraceAxisScore =>
+    typeof value === 'number' &&
+    Number.isInteger(value) &&
+    value >= 1 &&
+    value <= 5;
+
+/**
+ * Keeps only valid TRACE planner axes so downstream metadata stays schema-safe.
+ */
+const normalizePlannerTemperament = (
+    temperament: ResponseMetadataRuntimeContext['plannerTemperament']
+): ResponseMetadataRuntimeContext['plannerTemperament'] => {
+    if (!temperament) {
+        return undefined;
+    }
+
+    const normalized: NonNullable<
+        ResponseMetadataRuntimeContext['plannerTemperament']
+    > = {};
+    for (const axis of TRACE_AXIS_KEYS) {
+        const score = temperament[axis];
+        if (isTraceAxisScore(score)) {
+            normalized[axis] = score;
+        }
+    }
+
+    return Object.keys(normalized).length > 0 ? normalized : undefined;
+};
+
+const normalizePlannerReasonCode = (
+    status: ExecutionStatus,
+    reasonCode: ExecutionReasonCode | undefined
+): PlannerExecutionReasonCode | undefined => {
+    if (status === 'executed') {
+        return undefined;
+    }
+
+    if (
+        reasonCode === 'planner_runtime_error' ||
+        reasonCode === 'planner_invalid_output'
+    ) {
+        return reasonCode;
+    }
+
+    return 'planner_runtime_error';
+};
+
+const parsePlannerPurpose = (
+    value: unknown
+): PlannerExecutionPurpose | undefined =>
+    value === 'chat_orchestrator_action_selection' ? value : undefined;
+
+const parsePlannerContractType = (
+    value: unknown
+): PlannerExecutionContractType | undefined => {
+    if (
+        value === 'structured' ||
+        value === 'text_json' ||
+        value === 'fallback'
+    ) {
+        return value;
+    }
+
+    return undefined;
+};
+
+const parsePlannerApplyOutcome = (
+    value: unknown
+): PlannerExecutionApplyOutcome | undefined => {
+    if (
+        value === 'applied' ||
+        value === 'adjusted_by_policy' ||
+        value === 'not_applied'
+    ) {
+        return value;
+    }
+
+    return undefined;
+};
+
+const isSteerabilityControlId = (
+    value: unknown
+): value is SteerabilityControlId =>
+    value === 'workflow_mode' ||
+    value === 'evidence_strictness' ||
+    value === 'review_intensity' ||
+    value === 'provider_preference' ||
+    value === 'persona_tone_overlay' ||
+    value === 'tool_allowance';
+
+const normalizeEvaluatorReasonCode = (
+    status: ExecutionStatus,
+    reasonCode: ExecutionReasonCode | undefined
+): EvaluatorExecutionReasonCode | undefined => {
+    if (status === 'executed') {
+        return undefined;
+    }
+
+    if (reasonCode === 'evaluator_runtime_error') {
+        return reasonCode;
+    }
+
+    return 'evaluator_runtime_error';
+};
+
+const normalizeGenerationReasonCode = (
+    status: ExecutionStatus,
+    reasonCode: ExecutionReasonCode | undefined
+): GenerationExecutionReasonCode | undefined => {
+    if (status === 'executed') {
+        return undefined;
+    }
+
+    if (reasonCode === 'generation_runtime_error') {
+        return reasonCode;
+    }
+
+    return 'generation_runtime_error';
+};
+
+const normalizeToolReasonCode = (
+    status: ExecutionStatus,
+    reasonCode: ToolInvocationReasonCode | undefined
+): ToolInvocationReasonCode | undefined => {
+    if (
+        reasonCode === 'tool_not_requested' ||
+        reasonCode === 'tool_not_used' ||
+        reasonCode === 'search_rerouted_to_fallback_profile' ||
+        reasonCode === 'search_reroute_not_permitted_by_selection_source' ||
+        reasonCode === 'search_reroute_no_tool_capable_fallback_available' ||
+        reasonCode === 'tool_unavailable' ||
+        reasonCode === 'tool_execution_error' ||
+        reasonCode === 'tool_timeout' ||
+        reasonCode === 'tool_http_error' ||
+        reasonCode === 'tool_network_error' ||
+        reasonCode === 'tool_invalid_response' ||
+        reasonCode === 'search_not_supported_by_selected_profile' ||
+        reasonCode === 'unspecified_tool_outcome'
+    ) {
+        return reasonCode;
+    }
+
+    if (status === 'executed') {
+        return undefined;
+    }
+
+    return status === 'failed'
+        ? 'tool_execution_error'
+        : 'unspecified_tool_outcome';
+};
+
+/**
+ * Builds canonical ResponseMetadata for trace storage and UI rendering.
+ * All values are derived from control-plane context and API annotations.
+ *
+ * Semantics guardrail:
+ * - execution/workflow/workflowMode/steerabilityControls are structural records.
+ * - provenance/provenanceAssessment/chip scores may include deterministic
+ *   heuristic derivation when structural evidence is partial.
+ * - TRACE temperament is response posture metadata, separate from provenance.
+ */
+const buildResponseMetadata = (
+    assistantMetadata: AssistantResponseMetadata,
+    runtimeContext: ResponseMetadataRuntimeContext
+): ResponseMetadata => {
+    const responseId = crypto.randomBytes(6).toString('base64url').slice(0, 8);
+    const chainHash = crypto
+        .createHash('sha256')
+        .update(runtimeContext.conversationSnapshot)
+        .digest('hex')
+        .substring(0, 16);
+    const ninetyDaysMs = 90 * 24 * 60 * 60 * 1000;
+
+    const citations = Array.isArray(assistantMetadata.citations)
+        ? assistantMetadata.citations
+        : [];
+    const retrieval = runtimeContext.retrieval;
+    const classificationToolExecution = runtimeContext.executionContext?.tool;
+    const retrievalToolExecuted =
+        classificationToolExecution?.status === 'executed' &&
+        classificationToolExecution.toolName === 'web_search';
+    const provenanceClassification = classifyProvenanceWithSignals({
+        assistantProvenance: assistantMetadata.provenance,
+        citationCount: citations.length,
+        retrievalRequested: retrieval?.requested ?? false,
+        retrievalUsed: retrieval?.used ?? false,
+        retrievalToolExecuted,
+        workflowEvidence: runtimeContext.workflow !== undefined,
+        trustGraphEvidenceAvailable:
+            runtimeContext.trustGraphEvidenceAvailable ?? false,
+        trustGraphEvidenceUsed: runtimeContext.trustGraphEvidenceUsed ?? false,
+    });
+    // TODO(provenance-structural-first): Reduce heuristic dependence here by
+    // preferring explicit runtime retrieval/tool evidence signals wherever they
+    // are available and contract-stable.
+    const provenance = provenanceClassification.provenance;
+    const provenanceAssessment = provenanceClassification.assessment;
+    const tradeoffCount = resolveTradeoffCount(
+        assistantMetadata.tradeoffCount,
+        runtimeContext.plannerTemperament
+    );
+    const temperament = normalizePlannerTemperament(
+        runtimeContext.plannerTemperament
+    );
+    // TODO(trace-target-vs-final): If review-time runtime can revise TRACE
+    // posture, persist both target and final TRACE values. Preserve the
+    // invariant that workflow mode is routing metadata while TRACE remains
+    // answer-posture metadata.
+    const evidenceScore = isTraceAxisScore(assistantMetadata.evidenceScore)
+        ? assistantMetadata.evidenceScore
+        : undefined;
+    const freshnessScore = isTraceAxisScore(assistantMetadata.freshnessScore)
+        ? assistantMetadata.freshnessScore
+        : undefined;
+    const shouldDeriveRetrievedChips =
+        provenance === 'Retrieved' &&
+        (evidenceScore === undefined || freshnessScore === undefined);
+    const derivedRetrievedChips = shouldDeriveRetrievedChips
+        ? deriveRetrievedChips({
+              citationCount: citations.length,
+              intent: retrieval?.intent,
+              contextSize: retrieval?.contextSize,
+          })
+        : undefined;
+    const finalEvidenceScore =
+        evidenceScore ?? derivedRetrievedChips?.evidenceScore;
+    const finalFreshnessScore =
+        freshnessScore ?? derivedRetrievedChips?.freshnessScore;
+
+    if (
+        provenance === 'Retrieved' &&
+        (finalEvidenceScore === undefined || finalFreshnessScore === undefined)
+    ) {
+        logger.error(
+            'Retrieved response metadata is missing required evidence/freshness chips.',
+            {
+                responseId,
+                retrievalRequested: retrieval?.requested ?? false,
+                retrievalUsed: retrieval?.used ?? false,
+            }
+        );
+    }
+
+    const safetyTier: SafetyTier = 'Low';
+    const licenseContext = 'MIT + HL3';
+    const execution: ExecutionEvent[] = [];
+    const plannerExecution = runtimeContext.executionContext?.planner;
+    if (plannerExecution) {
+        // TODO(workflow-planner-step-metadata): Planner is not workflow-native
+        // in lineage yet. If workflows support multiple planner invocations,
+        // add explicit attempt/correlation metadata without letting planner
+        // events redefine orchestration authority or step ownership.
+        // Treat this block as metadata ingestion only, not a policy hook.
+        const normalizedPlannerReasonCode = normalizePlannerReasonCode(
+            plannerExecution.status,
+            plannerExecution.reasonCode
+        );
+        const validatedPlannerPurpose = parsePlannerPurpose(
+            plannerExecution.purpose
+        );
+        const validatedPlannerContractType = parsePlannerContractType(
+            plannerExecution.contractType
+        );
+        const validatedPlannerApplyOutcome = parsePlannerApplyOutcome(
+            plannerExecution.applyOutcome
+        );
+        const plannerExecutionSource =
+            'runtimeContext.executionContext.planner';
+        const hasValidMattered = typeof plannerExecution.mattered === 'boolean';
+        const sanitizedMatteredControlIds = Array.isArray(
+            plannerExecution.matteredControlIds
+        )
+            ? plannerExecution.matteredControlIds.filter(
+                  isSteerabilityControlId
+              )
+            : [];
+        const invalidPlannerFields: Array<{
+            field:
+                | 'purpose'
+                | 'contractType'
+                | 'applyOutcome'
+                | 'mattered'
+                | 'matteredControlIds';
+            value: unknown;
+        }> = [];
+        if (validatedPlannerPurpose === undefined) {
+            invalidPlannerFields.push({
+                field: 'purpose',
+                value: plannerExecution.purpose,
+            });
+        }
+        if (validatedPlannerContractType === undefined) {
+            invalidPlannerFields.push({
+                field: 'contractType',
+                value: plannerExecution.contractType,
+            });
+        }
+        if (validatedPlannerApplyOutcome === undefined) {
+            invalidPlannerFields.push({
+                field: 'applyOutcome',
+                value: plannerExecution.applyOutcome,
+            });
+        }
+        if (!hasValidMattered) {
+            invalidPlannerFields.push({
+                field: 'mattered',
+                value: plannerExecution.mattered,
+            });
+        }
+        if (!Array.isArray(plannerExecution.matteredControlIds)) {
+            invalidPlannerFields.push({
+                field: 'matteredControlIds',
+                value: plannerExecution.matteredControlIds,
+            });
+        }
+
+        if (
+            validatedPlannerPurpose === undefined ||
+            validatedPlannerContractType === undefined ||
+            validatedPlannerApplyOutcome === undefined ||
+            !hasValidMattered
+        ) {
+            logger.error(
+                'planner execution metadata dropped due invalid required fields',
+                {
+                    responseId,
+                    source: plannerExecutionSource,
+                    plannerStatus: plannerExecution.status,
+                    invalidPlannerFields,
+                }
+            );
+        } else {
+            execution.push({
+                kind: 'planner',
+                status: plannerExecution.status,
+                purpose: validatedPlannerPurpose,
+                contractType: validatedPlannerContractType,
+                applyOutcome: validatedPlannerApplyOutcome,
+                mattered: plannerExecution.mattered,
+                matteredControlIds: sanitizedMatteredControlIds,
+                profileId: plannerExecution.profileId,
+                ...(plannerExecution.originalProfileId !== undefined && {
+                    originalProfileId: plannerExecution.originalProfileId,
+                }),
+                ...(plannerExecution.effectiveProfileId !== undefined && {
+                    effectiveProfileId: plannerExecution.effectiveProfileId,
+                }),
+                provider: plannerExecution.provider,
+                model: plannerExecution.model,
+                ...(normalizedPlannerReasonCode !== undefined && {
+                    reasonCode: normalizedPlannerReasonCode,
+                }),
+                ...(plannerExecution.durationMs !== undefined && {
+                    durationMs: plannerExecution.durationMs,
+                }),
+            });
+        }
+    }
+    const evaluatorExecution = runtimeContext.executionContext?.evaluator;
+    if (evaluatorExecution) {
+        const normalizedEvaluatorReasonCode = normalizeEvaluatorReasonCode(
+            evaluatorExecution.status,
+            evaluatorExecution.reasonCode
+        );
+        execution.push({
+            kind: 'evaluator',
+            status: evaluatorExecution.status,
+            ...(evaluatorExecution.outcome !== undefined && {
+                evaluator: evaluatorExecution.outcome,
+            }),
+            ...(normalizedEvaluatorReasonCode !== undefined && {
+                reasonCode: normalizedEvaluatorReasonCode,
+            }),
+            ...(evaluatorExecution.durationMs !== undefined && {
+                durationMs: evaluatorExecution.durationMs,
+            }),
+        });
+    }
+    const toolExecution = runtimeContext.executionContext?.tool;
+    if (toolExecution) {
+        const normalizedToolReasonCode = normalizeToolReasonCode(
+            toolExecution.status,
+            toolExecution.reasonCode
+        );
+        execution.push({
+            kind: 'tool',
+            status: toolExecution.status,
+            toolName: toolExecution.toolName,
+            ...(normalizedToolReasonCode !== undefined && {
+                reasonCode: normalizedToolReasonCode,
+            }),
+            ...(toolExecution.durationMs !== undefined && {
+                durationMs: toolExecution.durationMs,
+            }),
+        });
+    }
+    const generationExecution = runtimeContext.executionContext?.generation;
+    if (generationExecution) {
+        const normalizedGenerationReasonCode = normalizeGenerationReasonCode(
+            generationExecution.status,
+            generationExecution.reasonCode
+        );
+        execution.push({
+            kind: 'generation',
+            status: generationExecution.status,
+            profileId: generationExecution.profileId,
+            ...(generationExecution.originalProfileId !== undefined && {
+                originalProfileId: generationExecution.originalProfileId,
+            }),
+            ...(generationExecution.effectiveProfileId !== undefined && {
+                effectiveProfileId: generationExecution.effectiveProfileId,
+            }),
+            provider: generationExecution.provider,
+            model: generationExecution.model,
+            ...(normalizedGenerationReasonCode !== undefined && {
+                reasonCode: normalizedGenerationReasonCode,
+            }),
+            ...(generationExecution.durationMs !== undefined && {
+                durationMs: generationExecution.durationMs,
+            }),
+        });
+    }
+    // TODO(workflow-execution-metadata): Extend execution events with lineage
+    // (id/parentId), timing (startedAt/finishedAt), and per-step usage/cost
+    // once multi-step workflow execution is enabled.
+    const generationEventModel = execution
+        .filter((event) => event.kind === 'generation')
+        .at(-1)?.model;
+
+    return {
+        responseId,
+        provenance,
+        safetyTier,
+        tradeoffCount,
+        chainHash,
+        licenseContext,
+        // TODO(workflow-execution-metadata): Remove modelVersion after metadata
+        // consumers migrate to execution[] as canonical timeline authority.
+        // Compatibility mirror for legacy consumers that still read only a
+        // single model string.
+        modelVersion:
+            generationEventModel ??
+            runtimeContext.modelVersion ??
+            runtimeConfig.openai.defaultModel,
+        staleAfter: new Date(Date.now() + ninetyDaysMs).toISOString(),
+        citations,
+        provenanceAssessment,
+        ...(runtimeContext.totalDurationMs !== undefined && {
+            totalDurationMs: runtimeContext.totalDurationMs,
+        }),
+        ...(execution.length > 0 && { execution }),
+        ...(runtimeContext.workflow !== undefined && {
+            workflow: runtimeContext.workflow,
+        }),
+        ...(runtimeContext.workflowMode !== undefined && {
+            workflowMode: runtimeContext.workflowMode,
+        }),
+        ...(runtimeContext.steerabilityControls !== undefined && {
+            steerabilityControls: runtimeContext.steerabilityControls,
+        }),
+        ...(evaluatorExecution?.outcome !== undefined && {
+            evaluator: evaluatorExecution.outcome,
+        }),
+        ...(temperament && { temperament }),
+        ...(finalEvidenceScore !== undefined && {
+            evidenceScore: finalEvidenceScore,
+        }),
+        ...(finalFreshnessScore !== undefined && {
+            freshnessScore: finalFreshnessScore,
+        }),
+    };
+};
+
+export { buildResponseMetadata };
